requirements.txt
# requirements.txt
discord.py
python-dotenv
google-generativeai
PyMuPDF
ebooklib
beautifulsoup4